+++ src/Service/PageListRepo.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace Mediawiki\Api\Service;
+
+use Mediawiki\Api\MediawikiApi;
+
+class PageListRepo {
+
+	/**
+	 * @var MediawikiApi
+	 */
+	private $api;
+
+	/**
+	 * @param MediawikiApi $api
+	 */
+	public function __construct( MediawikiApi $api ) {
+		$this->api = $api;
+	}
+
+	/**
+	 * @param string $name
+	 * @param bool|int $recursive layers of recursion to do
+	 * @returns array
+	 */
+	public function getPageListFromCategoryName( $name, $recursive = false ) {
+		//TODO implement me
+		throw new \BadMethodCallException( 'Not yet implemented' );
+	}
+
+}